requirements.txt
Flask==2.3.3
gunicorn==21.2.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.0.1
tensorflow==2.13.0
keras==2.13.1